Refactor. modify folder structure and import paths
