+++ src/app/App.jsx
@@ -1,0 +1,18 @@
+import { Routes, Route, Navigate } from "react-router-dom";
+
+import Login from "../components/pages/Login";
+import Header from "../components/Header";
+
+function App() {
+  return (
+    <>
+      <Header />
+      <Routes>
+        <Route path="/login" element={<Login />} />
+        <Route path="/" element={<Navigate replace to="/login" />} />
+      </Routes>
+    </>
+  );
+}
+
+export default App;
