+++ src/app/firebaseAuth.js
