--- src/components/app/App.jsx
@@ -1,13 +1,0 @@
-import { Routes, Route, Navigate } from "react-router-dom";
-import Login from "../pages/Login";
-
-function App() {
-  return (
-    <Routes>
-      <Route path="/login" element={<Login />} />
-      <Route path="/" element={<Navigate replace to="/login" />} />
-    </Routes>
-  );
-}
-
-export default App;
