+++ src/main.jsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
-import App from "./components/app/App";
+import App from "./app/App";
 import "./main.css";
 
 const queryClient = new QueryClient();
